make cache builder and actor which starts builder workers
cache builder will start and ask all of it's workers to build
