+++ lib/berkshelf/api/application.rb
@@ -11,8 +11,7 @@ module Berkshelf::API
     def initialize(registry)
       super
       supervise_as(:cache_manager, Berkshelf::API::CacheManager)
-      # @todo AG - Remove the :get_only
-      supervise_as(:cache_builder, Berkshelf::API::CacheBuilder::Opscode, get_only: 2)
+      supervise_as(:cache_builder, Berkshelf::API::CacheBuilder)
       supervise_as(:rest_gateway, Berkshelf::API::RESTGateway)
     end
   end
