+++ lib/berkshelf/api/cache_manager.rb
@@ -38,6 +38,7 @@ module Berkshelf::API
     attr_reader :cache
 
     def initialize
+      log.info "Cache Manager starting..."
       @cache = DependencyCache.new
       load_save if File.exist?(save_file)
     end
