--- spec/unit/berkshelf/api/cache_builder/chef_server_spec.rb
@@ -1,5 +1,0 @@
-require 'spec_helper'
-
-describe Berkshelf::API::CacheBuilder::ChefServer do
-  pending
-end
