+++ spec/unit/berkshelf/api/cache_builder/worker/chef_server_spec.rb
@@ -1,0 +1,5 @@
+require 'spec_helper'
+
+describe Berkshelf::API::CacheBuilder::Worker::ChefServer do
+  pending
+end
